Bitpanda/Challenge 2 Data Extraction/requirements.txt
streamlit
gspread
telethon
google-auth